Avoid the useless octree class construction

+++ src/Octree.h
@@ -59,7 +59,7 @@ private:
 	std::vector<std::vector<float>> m_sampleCache;
 
 	Octree* ConstructLeafParent(const int resolution, const glm::vec3 min);
-	Octree* ConstructLeaf(const int resolution, glm::vec3 min);
+	Octree* ConstructLeaf(const int resolution, glm::ivec3 min);
 
 	// Disallow constructors
 	Octree(const Octree&);
